+++ amd/__init__.py
@@ -1,21 +1,30 @@
 """
 average-minimum-distance: isometrically invariant crystal fingerprints
 ======================================================================
+
 Documentation is available in the docstrings and
 online at https://average-minimum-distance.readthedocs.io.
-Using functions from this package does not require explicit import of
-the relevant module.
+Using public functions from this amd does not require 
+explicit import of the module.
 
-::
+List of modules
+***************
 
- calculate   --- Calculate invariants (fingerprints) of periodic sets (crystals)
- compare     --- Compare fingerprints of crystals
- io          --- Read periodic sets from a file or the CSD
- periodicset --- Implements the PeriodicSet object
- pset_io     --- For writing/reading periodic sets + data with h5py
- reconstruct --- Implements recontructing a periodic set from PDD
- utils       --- General utility functions/classes (eg diameter, eta)
+===================    ===============================================================
+Module                 Description
+===================    ===============================================================
+:mod:`.calculate`      Calculate invariants (fingerprints) of periodic sets (crystals)
+:mod:`.compare`        Compare fingerprints of crystals
+:mod:`.io`             Read periodic sets from a file or the CSD
+:mod:`.periodicset`    Implements the PeriodicSet object
+:mod:`.utils`          General utility functions/classes (eg diameter, eta)
+===================    ===============================================================
+
 """
+
+__version__ = '1.2.2'
+__author__ = 'Daniel Widdowson'
+
 
 from .calculate import *
 from .compare import *